refactor(core): merge console-rcon package into runner
