--- console-rcon-package/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@plugwright/console-rcon",
-  "version": "3.0.0-dev.1",
-  "description": "RCON server console for plugwright's \"external\" mode",
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "rimraf dist && tsc",
-    "prepublishOnly": "npm run build",
-    "watch": "tsc --watch",
-    "typecheck": "tsc --noEmit"
-  },
-  "files": [
-    "dist"
-  ],
-  "keywords": [
-    "minecraft",
-    "rcon",
-    "plugwright",
-    "testing"
-  ],
-  "author": "drownek",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Drownek/plugwright.git",
-    "directory": "console-rcon-package"
-  },
-  "homepage": "https://github.com/Drownek/plugwright#readme",
-  "bugs": {
-    "url": "https://github.com/Drownek/plugwright/issues"
-  },
-  "peerDependencies": {
-    "@plugwright/runner": ">=3.0.0-dev.0"
-  },
-  "devDependencies": {
-    "@plugwright/runner": "file:../runner-package",
-    "@types/node": "^22.10.5",
-    "rimraf": "^6.1.3",
-    "typescript": "^5.7.3"
-  },
-  "engines": {
-    "node": ">=16.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
