+++ gradle-plugin/plugwright-local/src/main/kotlin/me/drownek/plugwright/local/LocalMode.kt
@@ -28,8 +28,7 @@ object LocalMode : PlugwrightMode<LocalEnvironmentSpec> {
         LocalEnvironmentSpec(name, objects)
 
     override fun runnerPackages(spec: LocalEnvironmentSpec): List<RunnerPackageRef> = listOf(
-        RunnerPackageRef("@plugwright/runner", export = "localEnvironment"),
-        RunnerPackageRef("@plugwright/console-rcon", export = "rconConsole"),
+        RunnerPackageRef("@plugwright/runner", export = "localEnvironment")
     )
 
     override fun validate(spec: LocalEnvironmentSpec, ctx: ValidationContext) {
