+++ runner-package/lib/environments/external.ts
@@ -6,7 +6,8 @@ import type { SecretRef } from '../config.js';
 import { resolveSecret } from '../config.js';
 import { AccountPool } from '../account.js';
 import type { AccountsConfig } from '../account.js';
-import { sleep, importOptionalPackage } from '../utils.js';
+import { sleep } from '../utils.js';
+import { rconConsole } from '../rcon/index.js';
 
 export interface ExternalConsoleChannelConfig {
     kind: 'rcon';
@@ -96,29 +97,7 @@ class ExternalEnvironment implements Environment {
         channel: ExternalConsoleChannelConfig,
     ): Promise<ServerConsole | null> {
         if (channel.kind === 'rcon') {
-            // A bare string literal here would make tsc try to resolve
-            // "@plugwright/console-rcon"'s types even though it's an optional peer package
-            // this repo doesn't depend on — routing through a variable keeps the import
-            // dynamic (untyped) without an ambient module declaration.
-            const rconPackage = '@plugwright/console-rcon';
-            let mod: any;
-            try {
-                mod = await importOptionalPackage(rconPackage);
-            } catch (error) {
-                console.error(pc.red(
-                    'Mode "external": console { rcon { } } needs the "@plugwright/console-rcon" package.\n' +
-                    'It installs automatically as part of plugwrightCompileTests — check that npm install\n' +
-                    'completed in your tests directory and that the package appears under node_modules.\n' +
-                    `(${(error as Error).message})`
-                ));
-                return null;
-            }
-            const factory = mod.rconConsole ?? mod.default;
-            if (typeof factory !== 'function') {
-                console.error(pc.red('"@plugwright/console-rcon" has no "rconConsole" export'));
-                return null;
-            }
-            return factory({
+            return rconConsole({
                 host: this.config.host,
                 port: channel.port ?? 25575,
                 password: channel.password ? resolveSecret(channel.password) : '',
